+++ src/main/java/com/service/BlogService.java
@@ -7,12 +7,12 @@ import com.dto.BlogDTO;
 public interface BlogService {
 	public List<BlogDTO> getAllBlogs();
 
-	public BlogDTO getBlogById(int blogId);
+	public BlogDTO getBlogById(Long blogId);
 
 	public BlogDTO createBlog(BlogDTO blogDto);
 
-	public BlogDTO updateBlog(int id, BlogDTO blogDTO);
+	public BlogDTO updateBlog(Long id, BlogDTO blogDTO);
 
-	public boolean deleteBlog(int id);
+	public boolean deleteBlog(Long id);
 
 }
